fixes, allow sats below 1

+++ interfaces/dvmtaskinterface.py
@@ -13,7 +13,7 @@ class DVMTaskInterface:
     NAME: str
     KIND: int
     TASK: str
-    COST: int
+    COST: float
     PRIVATE_KEY: str
     PUBLIC_KEY: str
     DVM = DVM
