+++ tasks/imagegeneration_openai_dalle.py
@@ -25,7 +25,7 @@ Params: -model         # models: juggernaut, dynavision, colossusProject, newrea
 class ImageGenerationDALLE(DVMTaskInterface):
     KIND: int = EventDefinitions.KIND_NIP90_GENERATE_IMAGE
     TASK: str = "text-to-image"
-    COST: int = 120
+    COST: float = 120
 
     def __init__(self, name, dvm_config: DVMConfig, nip89config: NIP89Config,
                  admin_config: AdminConfig = None, options=None):
